% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therParams.R
\name{therProject}
\alias{therProject}
\title{Project thermizer object}
\usage{
therProject(params)
}
\arguments{
\item{params}{A mizer params object}
}
\description{
Wrapper function adjusting simulation time and start time
for the project function
}
