% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{plankton_forcing}
\alias{plankton_forcing}
\title{plankton forcing}
\usage{
plankton_forcing(params, t, ...)
}
\arguments{
\item{params}{An object of class \linkS4class{MizerParams}.}

\item{t}{Time}

\item{...}{To pass further arguments down to the function}
}
\description{
Uses the values from the n_pp_array slot to produce
the resource spectrum.
}
