% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotThermPerformance}
\alias{plotThermPerformance}
\title{plotTherPerformance}
\usage{
plotThermPerformance(params, return_data = FALSE)
}
\arguments{
\item{params}{A mizer params object}
}
\description{
Take a look at the thermal performance of the species
}
