% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTherPerformance}
\alias{plotTherPerformance}
\title{plot thermal performance.}
\usage{
plotTherPerformance(params, resolution = 0.2, return_data = FALSE)
}
\arguments{
\item{params}{A mizer params object}

\item{resolution}{Numeric value which determines the step-width between each
calculation of the species' thermal performance. Default is 0.2.}

\item{return_data}{Boolean value allowing to return the data frame used for
the plot instead of the plot itself. Default is FALSE.}
}
\description{
Take a look at the thermal performance of the species.
}
