% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTherScalar}
\alias{plotTherScalar}
\title{plotTherScalar}
\usage{
plotTherScalar(
  params,
  species = NULL,
  species_panel = TRUE,
  return_data = FALSE
)
}
\arguments{
\item{params}{A mizer params object}

\item{species}{A character string. Select of specific species to display. It
has to correspond to one of the species name in the mizerParams object.
Default is NULL.}

\item{species_panel}{Boolean value. If set to TRUE, the plot will be a panel
of the species. Disabled if the argument species is not NULL. Default is TRUE.}

\item{return_data}{Boolean value allowing to return the data frame used for
the plot instead of the plot itself. Default is FALSE.}
}
\description{
Plot the scalar value affecting the encounter rate and metabolsim
of each species throughout the provided temperature in ocean_temp_array.
}
