% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_temp_effect.R
\name{scaled_temp_effect}
\alias{scaled_temp_effect}
\title{Temperature scaling factor}
\usage{
scaled_temp_effect(params, t)
}
\arguments{
\item{params}{An object of class \linkS4class{MizerParams}.}

\item{t}{Time}
}
\description{
Calculate the temperature scaling factor for the encounter rate and predation rate
}
