% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_temp_effect.R
\name{setEncounterPredScale}
\alias{setEncounterPredScale}
\title{Set encounter scalar}
\usage{
setEncounterPredScale(params)
}
\arguments{
\item{params}{An object of class \linkS4class{MizerParams}.}
}
\description{
Creates the encounterpred_scale parameter which is
used for scaling encounter and mortality rates and set the temperature scalar
between 0 and 1.
}
