% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_temp_effect.R
\name{setEncounterPredScale}
\alias{setEncounterPredScale}
\title{Encounter scale}
\usage{
setEncounterPredScale(params)
}
\arguments{
\item{params}{mizer object}
}
\description{
Create parameter for scaling encounter and mortality rates
}
