% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_temp_effect.R
\name{setMetabTher}
\alias{setMetabTher}
\title{Set metabolism temperature scaling parameters}
\usage{
setMetabTher(params)
}
\arguments{
\item{params}{A \code{MizerParams} object that has been prepared for
therMizer, typically with \code{\link{upgradeTherParams}()}.}
}
\value{
The modified \code{params} object with
  \code{species_params(params)$metab_min} and
  \code{species_params(params)$metab_range} filled in.
}
\description{
Compute the minimum and range of the Arrhenius-style metabolism response over
each species' thermal range. These values are later used to scale metabolic
effects between 0 and 1.
}
