% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_temp_effect.R
\name{setMetabTher}
\alias{setMetabTher}
\title{Metabolism temperature}
\usage{
setMetabTher(params)
}
\arguments{
\item{params}{An object of class \linkS4class{MizerParams}.}
}
\description{
Determine the minimum, maximum, and range of value for the
effect of temperature on metabolism.
}
