% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realms.R
\name{setVerticality}
\alias{setVerticality}
\title{set verticality}
\usage{
setVerticality(params, realm_names)
}
\arguments{
\item{params}{mizer params object}

\item{real_names}{A character vector containing the names of the
different realms to create.}
}
\description{
set different species specifc realms with different temperatures
}
\details{
Vertical integration functions
}
