% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therMizer-package.R
\docType{package}
\name{therMizer-package}
\alias{therMizer}
\alias{therMizer-package}
\title{therMizer: Extends the mizer package with temperature dynamics}
\description{
This is an extension package for the
\href{https://sizespectrum.org/mizer/}{mizer package}
that contains additional features regarding the effect
of temperature on organisms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sizespectrum/therMizer}
}

}
\author{
\strong{Maintainer}: Romain Forestier \email{romain.forestier@utas.edu.au} [copyright holder]

Authors:
\itemize{
  \item Romain Forestier \email{romain.forestier@utas.edu.au} [copyright holder]
  \item Phoebe Woodworth-Jefcoats \email{phoebe.woodworth-jefcoats@noaa.gov} (\href{https://orcid.org/0000-0001-9353-923X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
