% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therMizer-package.R
\docType{package}
\name{therMizer-package}
\alias{therMizer}
\alias{therMizer-package}
\title{therMizer: Extends the mizer package with temperature dynamics}
\description{
This is an extension package for the
\href{https://sizespectrum.org/mizer/}{mizer package}
that contains additional features regarding the effect
of temperature on organisms.
}
\keyword{internal}
