% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{therMizerEncounter}
\alias{therMizerEncounter}
\title{therMizerEncounter}
\usage{
therMizerEncounter(params, t, ...)
}
\arguments{
\item{params}{An object of class \linkS4class{MizerParams}.}

\item{t}{Time}

\item{...}{To pass further arguments down to the function}
}
\description{
Calculates the temperature-scaled encounter rate.
}
