% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{therMizerEReproAndGrowth}
\alias{therMizerEReproAndGrowth}
\title{therMizerEReproAndGrowth}
\usage{
therMizerEReproAndGrowth(params, t, encounter, feeding_level, ...)
}
\arguments{
\item{params}{An object of class \linkS4class{MizerParams}.}

\item{t}{Time}
}
\description{
Calculates the temperature-scaled energy available
for growth and reproduction.
}
