% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{therMizerEReproAndGrowth}
\alias{therMizerEReproAndGrowth}
\title{therMizererrepro}
\usage{
therMizerEReproAndGrowth(params, t, encounter, feeding_level, ...)
}
\arguments{
\item{params}{mizer object}

\item{t}{time}
}
\description{
bla
}
