% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{therMizerPredRate}
\alias{therMizerPredRate}
\title{therMizerPredRate}
\usage{
therMizerPredRate(params, n, n_pp, n_other, t, feeding_level, ...)
}
\arguments{
\item{params}{mizer object}

\item{t}{time}
}
\description{
bla
}
