% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therParams.R
\name{upgradeTherParams}
\alias{upgradeTherParams}
\title{Upgrade to thermizer object}
\usage{
upgradeTherParams(
  params,
  ocean_temp_array = NULL,
  n_pp_array = NULL,
  realm_names = NULL
)
}
\arguments{
\item{params}{A mizer params object}

\item{ocean_temp_array}{An array of temperatures}

\item{n_pp_array}{An array of plankton forcing}

\item{realm_names}{A character vector for realms names}
}
\description{
Wrapper function making a mizer params object
into something useable by the package
}
